+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 [< Back](./README.md)
 
-## Next
+## 0.1.3 (2026-03-06)
 
 ### Compiler